+++ packages/ui-core/shared/src/lib/project/project-mutation/project-mutation.component.ts
@@ -3,7 +3,6 @@ import { AbstractControl, UntypedFormBuilder, FormControl, UntypedFormGroup, Val
 import { EMPTY, firstValueFrom, of, switchMap } from 'rxjs';
 import { catchError, debounceTime, filter, finalize, tap } from 'rxjs/operators';
 import { NbDialogService } from '@nebular/theme';
-import { CKEditor4 } from 'ckeditor4-angular/ckeditor';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { TranslateService } from '@ngx-translate/core';
 import { Router } from '@angular/router';
@@ -47,7 +46,6 @@ import {
 import { TranslationBaseComponent } from '@gauzy/ui-core/i18n';
 import { patterns } from '../../regex/regex-patterns.const';
 import { FormHelpers } from '../../forms/helpers';
-import { ckEditorConfig } from '../../ckeditor.config';
 import { ProjectModuleMutationComponent } from '../../project-module/project-module-mutation/project-module-mutation.component';
 
 @UntilDestroy({ checkProperties: true })
@@ -71,7 +69,6 @@ export class ProjectMutationComponent extends TranslationBaseComponent implement
 	public owners: ProjectOwnerEnum[] = Object.values(ProjectOwnerEnum);
 	public taskViewModeTypes: TaskListTypeEnum[] = Object.values(TaskListTypeEnum);
 	public showSprintManage = false;
-	public ckConfig: CKEditor4.Config = ckEditorConfig;
 	public organization: IOrganization;
 	public hoverState: boolean;
 	public loading: boolean;