+++ packages/ui-core/shared/src/lib/rich-text-editor/index.ts
@@ -1,0 +1,2 @@
+export * from './rich-text-editor.module';
+export * from './rich-text-editor.component';